Remove dependency of `optimization/` tasks on `utils/training_loop.py`.
Instead, these tasks now return functions and processes that can serve as input to the training loop. This change allows more targeted unit tests, and clearly separates the API for the tasks from that of the training loop.

This change also introduces config classes for the input and output of these federated tasks, in order to consolidate documentation and reduce the size of callspaces.

PiperOrigin-RevId: 350586537
Change-Id: I544d07960d9e80ec629fa35ee26b61c0188794c5