Loop over the number of clients when sampling the labels.
This change fixes the loop sampling training/test examples for each client, looping over the number of clients, instead of the number of classes (both being equal by default). It also parameterizes the number of clients in load_cifar10_federated, and distinguishes the number of examples per label from the number of examples per client.

PiperOrigin-RevId: 361098722
Change-Id: I9f4aa8db85a0cfb6e688e6fbcdff7752b66dfe4a