Update `sql_client_data_utils` to use `tempfile` module instead of `os.environ`.
For context, the `sqlite` module does not support directly connecting to a database on a remote disk. During loading, we first check whether the database is on a remote disk. If so, the file will be first copied to a temporary directory, and then connected by `sqlite`. During saving, we first create a sqlite database locally at a temporary space, and then copy to the desired location.

PiperOrigin-RevId: 414632299
Change-Id: I42eca7e79ac317cfefc70771db4b71aac0c6de0e