Add hypcluster trainer.
This CL also adds an `initial_model_weights_list` argument to `build_hypcluster_train`.

We need to use `tff.aggregators.MeanFactory(no_nan_division=True)` in HypCluster training algorithm to avoid NaNs in the learned model, which can happen when a model is not selected by any client in a round.

PiperOrigin-RevId: 438374535
Change-Id: If7fb6225d55fc80cfbc4100ad0fcc39d7a7a94c9