Create a `personalization_benchmark` folder under /federated_research
It creates two subdirs: /cross_device and /cross_silo. This CL moves most of experimental p13n benchmark code to the /cross_device subdir. For the /cross_silo subdir, it adds a `README.md` file which points to an external github dir. This is because some of the cross-silo experiments are written by PyTorch, which makes it hard to co-locate under /federated_research.

Note that TedMulti is not included in the current CL. We will add it later after solving its dep on internal files.

PiperOrigin-RevId: 453814561
Change-Id: Icde77fa7fb8cbc2535fb68b54943011044acacbe